[refector] 코드리뷰 반영한 리팩토링

+++ src/components/MenuDot.tsx
@@ -44,14 +44,12 @@ const MenuDot: React.FC<MenuDotProps> = ({
 
   const handleConfirm = async () => {
     try {
-      if (index !== undefined && videoId) {
-        await deletePlaylist(videoId, () => {
-          if (deleteItem) {
+      if (index !== undefined && deleteItem) {
+        if (videoId) {
+          await deletePlaylist(videoId, () => {
             deleteItem(index);
-          }
-        });
-      } else {
-        if (index !== undefined && deleteItem) {
+          });
+        } else {
           deleteItem(index);
         }
       }
